pinctrl: stm32: improve debugfs information of pinconf-pins entry
This patch changes the way to register pin information in
stm32_desc_function struct in order to improve pinctrl debugfs information.
Indeed, "cat pinconf-pins" will indicate for a pin the number of the
alternate function selected but also (now) the "real" function of the pin.

Ex:
"pin 135 (PI7): alternate 10 (SAI2_FS_A) - ..."

NOT_UPSTREAMABLE as it is

Change-Id: I4f5abcd0acd6249efa0473407120b582563f5745
Signed-off-by: Alexandre Torgue <alexandre.torgue@st.com>