serial: stm32: defer probe for dma devices
Defer device probe operation when a DMA channel request probe deferral.

With this change both DMA channels are acquired before DMA channels are
configured, so that if one of the channel requests a probe deferral, no
received data will be lost. Once no probe deferral is expected, DMAs are
configured and any failure in their configuration will make the driver
to fallback to interrupt mode as prior this change.

Signed-off-by: Erwan Le Ray <erwan.leray@st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Change-Id: I2cd1529d3bab15812f7d93b716b329a9e9c82937
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/157273
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>