clk: stm32mp1: fix Kernel panic at boot when disabling sdmmc2
When we disable SDMMC1, SDMMC2, FMC, QSPI clocks we have to put
their MUX on save position.
In your case we are trying to set a MUX at save position to a bus
clock which have no MUX.
The bus clocks of SDMMC1, SDMMC2, QSPI and FMC are useless, we can
suppress it.
Use also flag instead a counter to know if we have to set the MUX
in safe position to manage the disable of clock ignore unused.

Change-Id: I7317e12d0d195ea7c9f36971d16c581901788a7a
Signed-off-by: Gabriel Fernandez <gabriel.fernandez@st.com>