dt-bindings: serial: stm32: support new rs485 delays in nanoseconds
Add rs485-rts-delay-ns (in nanoseconds) property.
This extends the range of delay values for user to microseconds and
nanoseconds.

With fast data rates (1Mbaut or more), 1ms become quite long. Users need
to be able to set shorter delays than 1 ms in order not to slow down the
channel unnecessarily. This enables faster communication channels without
changing the baudrate.

Signed-off-by: Erwan Le Ray <erwan.leray@st.com>
Change-Id: I9a1b8759440b9643760cfdc3d843a6b4a830a9f8