serial: stm32: update transmission complete error message in shutdown
The transmission complete error message provides the status of the
ISR_USART_TC bit. This bit, when set, indicates that the transmission
has not been completed.
The bit status indication is not a very understandable information.
The error message sent on console should indicate that the transmission is
not complete, instead of providing USART_TC bit status.
Update the error message and add a comment for better understanding.

Signed-off-by: Erwan Le Ray <erwan.leray@st.com>
Change-Id: Ie8abf83a7bf2f56c52a7a383976643441d0d92b3
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/188892
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>