serial: stm32: remove useless spin_unlock before calling dma_terminate
According to the comment in the code, the releasing of the lock around
dmaengine_terminate_all() is to avoid a deadlock with the DMA engine
callback.  However, since the time this code was written, the DMA engine
API documentation seems to have been clarified to say that
dmaengine_terminate_all() (in the identically implemented but
differently named dmaengine_terminate_async() variant) does not wait for
any running complete callback to be completed and can even be called
from a complete callback.  So there is no possibility of deadlock if the
DMA engine driver implements this API correctly.

See-also: f6a196477184 ("serial: pl011: Fix DMA ->flush_buffer()")

In stm32 driver, dma_terminate_all() have been removed and replace by
dma_terminate_async() in patch ("serial: stm32: update deprecated calls
to dmaengine_terminate_all") but not spin_lock() and spin_unlock().
There is no reason to keep them.

Signed-off-by: Valentin Caron <valentin.caron@foss.st.com>
Change-Id: I195fc6bc417818a09e2e980efde15ae9341371b1
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/194848
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Erwan LE RAY <erwan.leray@st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>